Set snyk tasks to never be up to date
